Missed a spot in r396559
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@396560 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ tests/test_hashtab_thrash.c
@@ -68,7 +68,7 @@ static int is_timed_out(struct hash_test const *data) {
 	if (val) {
 		/* tv_usec is suseconds_t, which could be int or long */
 		ast_test_status_update(data->test, "Now: %ld.%06ld Deadline: %ld.%06ld\n",
-			now.tv_sec, now.tv_usec,
+			now.tv_sec, (long)now.tv_usec,
 			data->deadline.tv_sec, (long)data->deadline.tv_usec);
 	}
 	return val;
